Numeric.FastMath.NaN

Description

This module contains rules that break the way NaN is handled for Float and Double types. Still, these rules should be safe in the vast majority of applications.

Importing this module is similar to compiling with gcc's -fno-signaling-nans and -ffinite-math-only.